refactor: extract function for finding import specifier
